PasswordGen
Creating a good password and recalling it is an expensive operation. You can create a program that takes certain words from the user and then creates a random password using those words. Users can remember the password with the help of the words they gave as input. Concluding our list of a handful of interesting ideas and projects you can build using Python, we can say that Python can be a very useful programming language to develop applications of all sorts and scales. If you have more suggestions or ideas, post it in the comment section, we’d love to hear about them